feat: implement a traceable class to put the common method
