+++ wren-ai-service/src/web/v1/services/__init__.py
@@ -1,0 +1,14 @@
+class MetadataTraceable:
+    def with_metadata(self) -> dict:
+        return {
+            "resource": self,
+            "metadata": {
+                **self._error_metadata(),
+            },
+        }
+
+    def _error_metadata(self):
+        return {
+            "error_type": self.error and self.error.code,
+            "error_message": self.error and self.error.message,
+        }
